+++ bit/count.go
@@ -17,12 +17,12 @@ const (
 	ms32 = m32 & b32
 )
 
-func Count(x uint64) uint64 {
+func Count(x uint64) uint {
 	x = (x & m1) + ((x >> 1) & m1)
 	x = (x & m2) + ((x >> 2) & m2)
 	x = (x & m4) + ((x >> 4) & m4)
 	x = (x & m8) + ((x >> 8) & m8)
 	x = (x & m16) + ((x >> 16) & m16)
 	x = (x & m32) + ((x >> 32) & m32)
-	return uint64(x)
+	return uint(x)
 }
